browser and node.js use different mounting for the library
the webpack config files for ESM and CJS need different build flags.

CommonJS modules should be using "this" rather than "self" to install
the library. Without that change, the require() call fails with
"self is undefined" error.

When using the ESM module within a browser, "self" is used during
the import of the module.